+++ CHANGELOG.md
@@ -24,6 +24,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 ## [Unreleased]
 
 * Fix issue #548: Make time coordinate regularity check less stringent.
+* Force `DataTensors` to have all tensors with `dtype=tf.float32`.
 
 ## [1.0.5] - 2025-03-06